Add investor page with email investor@21scenturyprogram.com
We're adding this page for an outreach email to One Mind Accelerator. One mind is a foundation focused on brain health research and development based in the US. They started a new accelerator and we were curious about it. 

You can find One Mind on the web at onemind.org